requirements.txt
discord.py
chat-exporter
PyGithub
aiosqlite